requirements.txt
# Парсинг
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0

# Дедупликация
rapidfuzz>=3.5.0

# БД
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Веб-приложение
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0

# Утилиты
python-dotenv>=1.0.0
